tu/tu/requirements.txt
Django
djangorestframework
requests
mysqlclient
